+++ src/mjolnir/graphtilebuilder.cc
@@ -584,9 +584,28 @@ void GraphTileBuilder::AddSigns(const uint32_t idx, const std::vector<SignInfo>&
 }
 
 // Add lane connectivity
-void GraphTileBuilder::AddLaneConnectivity(const std::vector<baldr::LaneConnectivity>& lc) {
-  lane_connectivity_builder_.insert(lane_connectivity_builder_.end(), lc.begin(), lc.end());
-  lane_connectivity_offset_ += sizeof(baldr::LaneConnectivity) * lc.size();
+void GraphTileBuilder::AddLaneConnectivity(std::vector<baldr::LaneConnectivity>&& lc) {
+  size_t size = lc.size();
+
+  lane_connectivity_builder_.reserve(lane_connectivity_builder_.size() + size);
+  lane_connectivity_builder_.insert(lane_connectivity_builder_.end(),
+                                    std::make_move_iterator(lc.begin()),
+                                    std::make_move_iterator(lc.end()));
+  lane_connectivity_offset_ += sizeof(baldr::LaneConnectivity) * size;
+}
+
+void GraphTileBuilder::CopyLaneConnectivityFromTile(const baldr::graph_tile_ptr& tile,
+                                                    uint32_t edge_id) {
+  auto laneconnectivity_span = tile->GetLaneConnectivity(edge_id);
+  auto laneconnectivity = std::vector<baldr::LaneConnectivity>(laneconnectivity_span.begin(),
+                                                               laneconnectivity_span.end());
+  if (laneconnectivity.size() == 0) {
+    LOG_ERROR("Base edge should have lane connectivity, but none found");
+  }
+  for (auto& lc : laneconnectivity) {
+    lc.set_to(directededges().size());
+  }
+  AddLaneConnectivity(std::move(laneconnectivity));
 }
 
 // Add forward complex restriction.